pyproject.toml
[project]
name = "pars-diary"
version = "0.1.1"
description = "Проект для улучшения успеваемости учеников, путем внедрения современных технологий."
authors = [{ name = "Ivan Melcikov", email = "vanamelcikov7275@gmail.com" }]
license = "MIT"
readme = "README.md"

dependencies = [
    "matplotlib>=3.10.0, <4.0",
    "gradio_client>=1.3.0, <2.0",
    "loguru>=0.7.3, <0.8",
    "aiogram>=3.19.0, <4.0",
    "python-dotenv>=1.1.0, < 2.0",
    "conversations-about-important-api==0.1.0",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Ошибки связанные с Кириллицей
    "RUF001",
    "RUF002",
    "RUF003",

    # We use modern python
    "FA",
    "COM812",
    "D203",
    "D213",

    # TODO (need issue link; )
    "TD003",
    "FIX002",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]
